+++ BambooTracker/playback.cpp
@@ -535,6 +535,7 @@ bool PlaybackManager::setEffectToQueueFM(int ch, Effect eff)
 	case EffectType::TransposeDelay:
 	case EffectType::VolumeDelay:
 	case EffectType::FBControl:
+	case EffectType::TLControl:
 		keyOnBasedEffsFM_.at(static_cast<size_t>(ch)).push_back(std::move(eff));
 		break;
 	case EffectType::SpeedTempoChange:
@@ -661,13 +662,20 @@ bool PlaybackManager::readFMEffectFromQueue(int ch)
 				int count = eff.value >> 8;
 				if (count > 0) {
 					volDlyCntFM_.at(uch) = count;
-					volDlyValueFM_.at(uch) = eff.value;
+					volDlyValueFM_.at(uch) = eff.value & 0x00ff;
 				}
 				break;
 			}
 			case EffectType::FBControl:
 				if (-1 < eff.value && eff.value < 8) opnaCtrl_->setFBControlFM(ch, eff.value);
 				break;
+			case EffectType::TLControl:
+			{
+				int op = eff.value >> 8;
+				int val = eff.value & 0x00ff;
+				if (-1 < op && op < 4 && -1 < val && val < 128) opnaCtrl_->setTLControlFM(ch, op, val);
+				break;
+			}
 			default:
 				break;
 			}
@@ -834,7 +842,7 @@ bool PlaybackManager::readSSGEffectFromQueue(int ch)
 				int count = eff.value >> 8;
 				if (count > 0) {
 					volDlyCntSSG_.at(uch) = count;
-					volDlyValueSSG_.at(uch) = eff.value;
+					volDlyValueSSG_.at(uch) = eff.value & 0x00ff;
 				}
 				break;
 			}
@@ -948,7 +956,7 @@ bool PlaybackManager::readDrumEffectFromQueue(int ch)
 				int count = eff.value >> 8;
 				if (count > 0) {
 					volDlyCntDrum_.at(uch) = count;
-					volDlyValueDrum_.at(uch) = eff.value;
+					volDlyValueDrum_.at(uch) = eff.value & 0x00ff;
 				}
 				break;
 			}
@@ -1243,7 +1251,7 @@ void PlaybackManager::retrieveChannelStates()
 	std::vector<bool> isSetInstFM(fmch, false), isSetVolFM(fmch, false), isSetArpFM(fmch, false);
 	std::vector<bool> isSetPrtFM(fmch, false), isSetVibFM(fmch, false), isSetTreFM(fmch, false);
 	std::vector<bool> isSetPanFM(fmch, false), isSetVolSldFM(fmch, false), isSetDtnFM(fmch, false);
-	std::vector<bool> isSetFBCtrlFM(fmch, false);
+	std::vector<bool> isSetFBCtrlFM(fmch, false), isSetTLCtrlFM(fmch, false);
 	std::vector<bool> isSetInstSSG(3, false), isSetVolSSG(3, false), isSetArpSSG(3, false), isSetPrtSSG(3, false);
 	std::vector<bool> isSetVibSSG(3, false), isSetTreSSG(3, false), isSetVolSldSSG(3, false), isSetDtnSSG(3, false);
 	std::vector<bool> isSetTNMixSSG(3, false);
@@ -1360,6 +1368,18 @@ void PlaybackManager::retrieveChannelStates()
 								if (-1 < eff.value && eff.value < 8) opnaCtrl_->setFBControlFM(ch, eff.value);
 							}
 						}
+						break;
+					case EffectType::TLControl:
+						if (!isSetTLCtrlFM[uch]) {
+							isSetTLCtrlFM[uch] = true;
+							if (isPrevPos) {
+								int op = eff.value >> 8;
+								int val = eff.value & 0x00ff;
+								if (-1 < op && op < 4 && -1 < val && val < 128)
+									opnaCtrl_->setTLControlFM(ch, op, val);
+							}
+						}
+						break;
 					default:
 						break;
 					}